backend/requirements.txt
boto3==1.28.63
google-auth==2.22.0
PyJWT==2.8.0
cryptography==41.0.4
requests==2.31.0 